Address review feedback, add format version to materialized view properties
